fix at line 43 for type in modulo

+++ encoder_master/EncoderBoard.cpp
@@ -18,29 +18,29 @@ EncoderBoard::EncoderBoard(int address) {
 //////////////////////////////////////////////////////////////////////////////////
 
 void EncoderBoard::update() {
-  Wire.requestFrom(_address, sizeof(_data_u));    // request 128 bytes from slave device 
+  Wire.requestFrom(_address, sizeof(_data_u));    // request 128 bytes from slave device
   int i = 0;
   while(Wire.available()) { // slave may send less than requested
-    _data_u.data_byte[i] = (Wire.read()); 
+    _data_u.data_byte[i] = (Wire.read());
     i++;
   }
 }
 
 //////////////////////////////////////////////////////////////////////////////////
 
-double EncoderBoard::get(int index) {	
+double EncoderBoard::get(int index) {
   return modulo(_data_u.data.angles[index] / COM_MULT ) + 2 * PI * _data_u.data.rounds[index];
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////
 
-double EncoderBoard::angular_speed(int index) {	
+double EncoderBoard::angular_speed(int index) {
   return _data_u.data.angular_speed[index] / COM_MULT_SPEED;
 }
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////
- 
-double EncoderBoard::modulo(int value) {	
+
+double EncoderBoard::modulo(double value) {	
   if(value < 0) return value;
   else return -value;
 }
